Updated mc_hms call in mc_single_arm.f
The mc_hms call was passing fry, the vertical
beam position, for the
vertical target position.

It should pass the xtar_init which is
tthe vertical target position at the z=0 plane
which includes fry.

The xtar_init is used for the xtar dependent
corrections in the recon matrix.

Ideally one would calculate the xtar and the recon matrix in
an iterative approach to mimic what  is done in HCANA.